building out hubs, projects, and a common file for the dm package
