+++ dm/projects.go
@@ -1,12 +1,12 @@
 package dm
 
-import (
-	"bytes"
-	"encoding/json"
-	"net/http"
-	"strconv"
-	"github.com/outer-labs/forge-api-go-client/oauth"
-)
+// import (
+// 	"bytes"
+// 	"encoding/json"
+// 	"net/http"
+// 	"strconv"
+// 	"github.com/outer-labs/forge-api-go-client/oauth"
+// )
 
 type ProjectDetails struct {
 	Data    []Content `json:"data, omitempty"`
@@ -14,35 +14,13 @@ type ProjectDetails struct {
 	Links   Link      `json:"links, omitempty"`
 }
 
-type JsonAPI struct {
-	Version string `json:"version, omitempty"`
-}
-
-type Link struct {
-	Self struct {
-		Href 	string `json:"href, omitempty"`
-	} `json:"self, omitempty"`
-	First struct {
-		Href string `json:"href, omitempty"`
-	} `json:"first, omitempty"`
-	Prev struct {
-		Href string `json:"href, omitempty"`
-	} `json:"prev, omitempty"`
-	Next struct {
-		Href string `json:"href, omitempty"`
-	} `json:"next, omitempty"`
-	Related struct {
-		Href string `json:"href, omitempty"`
-	} `json:"related, omitempty"`
-}
-
-type Content struct {
-	Relationships Relationships `json:"relationships, omitempty"`
-	Attributes Attribute `json:"attributes, omitempty"`
-	Type       string    `json:"type, omitempty"`
-	Id         string    `json:"id, omitempty"`
-	Links      Link      `json:"links, omitempty"`
-}
+// type Content struct {
+// 	Relationships Relationships `json:"relationships, omitempty"`
+// 	Attributes Attribute `json:"attributes, omitempty"`
+// 	Type       string    `json:"type, omitempty"`
+// 	Id         string    `json:"id, omitempty"`
+// 	Links      Link      `json:"links, omitempty"`
+// }
 
 type Relationships struct {
 	Hub 		[]Hub 		`json:"hub, omitempty"`
@@ -70,14 +48,3 @@ type Project struct {
 	Links Link `json:"links, omitempty"`
 }
 
-type Attribute struct {
-	Name      string `json:"name, omitempty"`
-	Extension struct {
-		Data    map[string]interface{} `json:"data, omitempty"`
-		Version string                 `json:"version, omitempty"`
-		Type    string                 `json:"type, omitempty"`
-		Schema  struct {
-			Href string `json:"href, omitempty"`
-		} `json:"schema, omitempty"`
-	} `json:"extension, omitempty"`
-}
